requirements.txt
# PyTorch 是一个深度学习框架，用于构建和训练神经网络。
# 这里要求 PyTorch 版本为 1.9.0 或更高版本，确保你可以使用最新的深度学习特性和性能改进。
torch>=1.9.0

# torchvision 是 PyTorch 的一个附加库，提供了图像处理的工具和预训练的模型。
# 该版本要求为 0.10.0 或更高，支持最新的图像处理操作和预训练模型。
torchvision>=0.10.0

# Flask 是一个轻量级的 Web 框架，用于构建 Web 应用程序。
# 这里我们使用 Flask 来构建一个 Web 服务接口，接收用户的图像输入并返回预测结果。
flask

# numpy 是一个强大的数学库，提供了高效的数组操作和线性代数运算。
# 它在处理图像数据、计算损失函数等方面非常重要。
numpy

# matplotlib 是一个用于创建静态、动态和交互式图形的 Python 库。
# 在深度学习中，我们常用它来可视化训练过程、损失变化、预测结果等。
matplotlib

# scikit-learn 是一个机器学习库，包含了大量的机器学习算法以及评估工具。
# 在项目中，可能会使用它来做数据预处理、数据增强等操作。
scikit-learn

# Pillow（PIL 的一个分支）是用于图像处理的库，支持打开、操作和保存图像。
# 在数据预处理中，通常使用 Pillow 读取和操作图像文件。
# Pillow 是用于图像处理的库，支持打开、操作和保存图像。
# 它是 PIL（Python Imaging Library）的一个分支，推荐使用 Pillow 代替 PIL。
Pillow

# opencv-python 是 OpenCV 库的 Python 接口，用于图像和视频处理。
# 它通常用于更复杂的图像处理任务，例如图像增强、边缘检测等。
opencv-python

# tqdm 是一个用于显示进度条的 Python 库，尤其在长时间运行的任务中，能帮助用户跟踪进度。
# 在训练过程中，我们会用它来显示训练进度条。
tqdm
